Updated ArticleController to use \Throwable; fixed typo in addArticle.blade.php; added admin user management routes
The ArticleController has been updated to use the \Throwable exception class instead of Throwable in the catch blocks of the store, update, and destroy methods. In the show method, a typo in the addArticle.blade.php file has been fixed where $$articleTags was changed to $articleTags in the foreach loop. Lastly, new routes have been added for admin user management in the web.php file.